requirements.txt
Flask==3.1.0
flask-sqlalchemy==3.1.1
numpy==2.1.3
pandas==2.2.3
scipy==1.14.1
openpyxl==3.1.5
pymysql==1.1.1
python-dotenv==1.0.1